+++ bot_samek.py
@@ -1,18 +1,23 @@
 import argparse
 import os
+import random
+from datetime import datetime, timedelta
 
-from loguru import logger
+SEED = random.randint(0, 1000)
 
 
-def flix_dou_yah(dataset, ith, seed, pe_dim, scenario):
+def flix_dou_yah(dataset, pe_dim, scenario):
+    now = datetime.now().strftime("%b%d_%H:%M")
+
     os.system(
         f"python train.py --data-name {dataset}\
-            --wandb -g 0 --version 2\
-            --save-appendix _gatedGCN_{pe_dim}_{scenario}_{ith}\
-            --data-appendix _pyg2.0_{pe_dim}\
-            --seed {seed}\
+            -g 0 --version 2\
+            --save-appendix _gatedGCN_{pe_dim}_{scenario}\
+            --data-appendix _{pe_dim}\
+            --seed {SEED}\
             --pe-dim {pe_dim}\
-            --scenario {scenario}"
+            --scenario {scenario} > logs/{dataset}_{scenario}_{pe_dim}_{now}.log"
+        # --wandb
     )
 
 
@@ -28,12 +33,7 @@ if __name__ == "__main__":
         "flixster": 20,
     }
 
-    seeds = [37, 10, 4, 73, 21]
-
     for dataset in ["yahoo_music", "douban", "flixster"]:
-        for ith, seed_val in enumerate(seeds):
-            logger.info(f"Test: {ith} - {dataset} - seed: {seed_val:3d}")
-
-            flix_dou_yah(dataset, ith, seed_val, pe_dims[dataset], args.scenario)
+        flix_dou_yah(dataset, pe_dims[dataset], args.scenario)
 
     # python Main.py --data-name yahoo_music --epochs 40 --testing --ensemble --save-appendix _80 --data-appendix _80 --seed 42 --pe-dim 80